Tarefa2_2019li1g102

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testJogadores :: [Jogador]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

Variaveis de teste para '[Jogador]' Mapa Estado

testMapa :: Mapa

testEstado :: Estado

testesT2 :: [(Int, Jogada, Estado)]

Funções principais da Tarefa 2.

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

Funções Auxiliares

Funções Gerais para as jogadas

estaMorto :: Jogador -> Bool

Retorna True caso o jogador esteja Morto e False caso contrário

estaNoChao :: Jogador -> Bool

Retorna True caso o jogador esteja no Chao e False caso contrário

estaNoAr :: Jogador -> Bool

Retorna True caso o jogador esteja no Ar e False caso contrário

mataJogador :: Jogador -> Jogador

Função que atualiza o Estado de um Jogador para Morto com um timeout de 1.0

mostraJogador :: Int -> Estado -> Jogador

Retorna um Jogador identificado pelo seu índice

coordenadaJogador :: Jogador -> (Int, Int)

Retorna os indices da matriz da posição atual do Jogador

atualiza :: Int -> Jogador -> [Jogador] -> [Jogador]

Atualiza o Jogador identificado por um indice na '[Jogador]'

Funções de Acelera e Desacelera

funAcelera :: Int -> Estado -> Estado

Atualiza o Estado do jogo após um Jogador acelerar

funDesacelera :: Int -> Estado -> Estado

Atualiza o Estado do jogo após um Jogador desacelerar

atualizaAcelaracao :: Jogador -> Jogador

Retorna um jogador com o EstadoJogador = Chao True

atualizaDesacelaracao :: Jogador -> Jogador

Retorna um Jogador com o EstadoJogador = Chao False

Funções de Direcao

funDirecao :: Int -> Estado -> Direcao -> Estado

Atualiza o Estado do jogo após um Jogador mudar de Direcao

getEstadoAr :: Jogador -> (Double, Double)

Retorna a altura e a inclinacao de um jogador

atualizaDirecao :: Jogador -> Direcao -> Mapa -> Jogador

Função que dado um Jogador e uma Direcao atualiza a sua posição no Mapa

altInc :: Jogador -> Mapa -> (Double, Double)

Retorna um (Double,Double) com a altura e a inclinaçao a que o Jogador se encontra

altIncSeguinte :: Jogador -> Mapa -> Direcao -> (Double, Double)

Retorna um (Double,Double) com a altura e a inclinaçao da Peca para qual o Jogador se quer mudar

alturasPeca :: Peca -> (Double, Double)

Retorna um (Double,Double) com a altura inicial e final de uma Peca respetivamente

movePista :: Jogador -> Direcao -> Int -> Jogador

Funcao que move o Jogador 1 pista para C(cima) ou para B(baixo)

atualizaAltInc :: Jogador -> Double -> Double -> Jogador

Altera o EstadoJogador para refletir a nova altura e inclinação do Jogador

Funções de Dispara

funDispara :: Int -> Estado -> Estado

Função que faz o Jogador identificado pelo seu indice disparar uma Cola

municao :: Jogador -> Int

Retorna a quantidade de munições que um Jogador tem

consomeMunicao :: Jogador -> Jogador

Retorna o Jogador apos consumir uma municao de Cola

funToCola :: Peca -> Peca

Altera o Piso de uma Peca para Cola