Tarefa4_2019li1g102

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [(Double, Mapa, Jogador)]

Testes unitários da Tarefa 4.

Cada teste é um par (tempo,Mapa,Jogador).

j0 :: Jogador

j1 :: Jogador

j2 :: Jogador

j3 :: Jogador

j4 :: Jogador

j5 :: Jogador

j6 :: Jogador

j7 :: Jogador

j8 :: Jogador

j10 :: Jogador

j11 :: Jogador

j12 :: Jogador

j13 :: Jogador

Funções principais da Tarefa 4.

passo

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após um passo.

Avança o estado de um Jogador um passo em frente, durante um determinado período de tempo.

acelera

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após acelerar.

Altera a velocidade de um Jogador, durante um determinado período de tempo.

move

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após se movimentar.

Altera a posição de Jogador, durante um determinado período de tempo.

Funções auxiliares

novaVelocidade :: Double -> Mapa -> Jogador -> Double

Funçao que calcula uma velocidade depois de um jogador acelarar no chão (v' = v + (accelMota - atrito * v) * t) ou ,caso esteja no ar, (v' = v - (resistenciaAr * v * t))

arredondaNegativos :: Double -> Double

Retorna o Double que recebe ou 0 caso seja menor que 0

accelMota :: Jogador -> Double

Retorna 1 caso seja possivel acelerar e 0 caso contrário

accelJogador :: Jogador -> Bool

Retorna True caso o Jogador esteja no Chao a acelerar e False caso contrário

getPiso :: Peca -> Piso

Retorna o Piso de uma Peca

atrito :: Peca -> Double

Retorna o atrito de cada Peca

novaGravidade :: Jogador -> Double -> Jogador

Nova Gravidade do jogador

fromMorto :: Double -> Jogador -> Jogador

Retorna um Jogador após estar morto durante t segundos

fromChao :: Double -> Mapa -> Jogador -> Jogador

Retorna um Jogador inicialmente no chao passados t segundos

fromAr :: Double -> Mapa -> Jogador -> Jogador

Retorna um Jogador inicialmente no ar passados t segundos

inclinacaoPeca :: Peca -> Double

Função que calcula a inclinação de uma Peca

altPeca :: Peca -> Double

Função que retorna a altura final de uma Peca

vetorVel :: Jogador -> Mapa -> Vetor

Retorna o Vetor da velocidade de um Jogador

vetorGrav :: Jogador -> Vetor

Retorna o Vetor da gravidade de um Jogador

retaPeca :: Double -> Peca -> Reta

Função que dá a Reta com o declive da Peca correspondente

retaVerticalPeca :: Jogador -> Reta

Função que calcula a Reta Vertical do fim da peça em que o jogador se encontra